Todo Letest 2/Todo/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
#backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pycodestyle==2.8.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
whitenoise==6.2.0
wheel==0.37.1
